Move HeapSize to heap_size module
